// apps/web/src/api/model/messageResponseDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Future Media Challenge API
 * The API for the Future Media Challenge application
 * OpenAPI spec version: 1.0
 */
import type { MessageAuthorDto } from './messageAuthorDto';
import type { MessageTagDto } from './messageTagDto';

export interface MessageResponseDto {
  id: string;
  /** @maxLength 240 */
  content: string;
  author: MessageAuthorDto;
  tag: MessageTagDto;
  createdAt: string;
  updatedAt: string;
}